problem 2
 c language

+++ C/Easy/2.c
@@ -1,7 +1,7 @@
 int main() {
     int a = 10;
     double b = 5.5;
-    int c = a + b;
-    printf("The sum of a and b is %d", c);
+    double c = a + b;
+    printf("The sum of a and b is %lf", c);
     return 0;
 }
